Add `rows_in_batches` method to PostgreSQLCursor
`instance_in_batches` is great for use with an ActiveRecord model, but
there was no way to submit a SQL string and get batches of rows. This
refactors `instance_in_batches` to use a method `rows_in_batches`
internally and expose `rows_in_batches` as a public method for direct
consumption.